+++ state/src/test/java/io/github/astrapi69/design/pattern/state/wizard/model/WizardModel.java
@@ -1,0 +1,72 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.design.pattern.state.wizard.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+/**
+ * The class {@link WizardModel} act as a model for a wizard.<br>
+ * This wizard model serves only as an example.<br>
+ * You can create a specific wizard model to your requirements.
+ */
+@Getter
+@Setter
+@EqualsAndHashCode
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder(toBuilder = true)
+public class WizardModel
+{
+
+	/** The flag that signals if next is valid or not. */
+	private boolean validNext;
+
+	/** The flag that signals if previous is valid or not. */
+	private boolean validPrevious;
+
+	/** The flag that signals if cancel is valid or not. */
+	private boolean validCancel;
+
+	/** The flag that signals if finish is valid or not. */
+	private boolean validFinish;
+
+	/**
+	 * Reset all flags to false.
+	 */
+	public void reset()
+	{
+		validNext = false;
+		validPrevious = false;
+		validCancel = false;
+		validFinish = false;
+	}
+}